@@ -8,14 +8,14 @@ const OS = require('os');
 const debug = require('debug')('log4js:gelf');
 
 /* eslint no-unused-vars:0 */
-const LOG_EMERG = 0;    // system is unusable(unused)
-const LOG_ALERT = 1;    // action must be taken immediately(unused)
-const LOG_CRIT = 2;     // critical conditions
-const LOG_ERROR = 3;    // error conditions
-const LOG_WARNING = 4;  // warning conditions
-const LOG_NOTICE = 5;   // normal, but significant, condition(unused)
-const LOG_INFO = 6;     // informational message
-const LOG_DEBUG = 7;    // debug-level message
+const LOG_EMERG = 0; // system is unusable(unused)
+const LOG_ALERT = 1; // action must be taken immediately(unused)
+const LOG_CRIT = 2; // critical conditions
+const LOG_ERROR = 3; // error conditions
+const LOG_WARNING = 4; // warning conditions
+const LOG_NOTICE = 5; // normal, but significant, condition(unused)
+const LOG_INFO = 6; // informational message
+const LOG_DEBUG = 7; // debug-level message
 
 /**
  * GELF appender that supports sending UDP packets to a GELF compatible server such as Graylog
@@ -113,7 +113,7 @@ function gelfAppender(layout, config, levels) {
 
   const app = (loggingEvent) => {
     const message = preparePacket(loggingEvent);
-    zlib.gzip(new Buffer(JSON.stringify(message)), (err, packet) => {
+    zlib.gzip(Buffer.from(JSON.stringify(message)), (err, packet) => {
       if (err) {
         console.error(err.stack);
       } else {
