@@ -21,7 +21,7 @@ function datagram(config) {
   const port = config.port || 55201;
 
   return function (event) {
-    const buff = new Buffer(JSON.stringify(event));
+    const buff = Buffer.from(JSON.stringify(event));
     sock.send(buff, 0, buff.length, port, host, (err) => {
       if (err) {
         console.error(`log4js.logFacesUDPAppender error sending to ${host}:${port}, error: `, err);
@@ -46,11 +46,11 @@ function logFacesUDPAppender(config) {
   return function log(event) {
     // convert to logFaces compact json format
     const lfsEvent = {
-      a: config.application || '',   // application name
-      t: event.startTime.getTime(),  // time stamp
-      p: event.level.levelStr,       // level (priority)
-      g: event.categoryName,         // logger name
-      m: format(event.data)          // message text
+      a: config.application || '', // application name
+      t: event.startTime.getTime(), // time stamp
+      p: event.level.levelStr, // level (priority)
+      g: event.categoryName, // logger name
+      m: format(event.data) // message text
     };
 
     // add context variables if exist
@@ -82,8 +82,9 @@ function wrapErrorsWithInspect(items) {
 }
 
 function format(logData) {
-  const data = Array.isArray(logData) ?
-               logData : Array.prototype.slice.call(arguments);
+  const data = Array.isArray(logData)
+    ? logData
+    : Array.prototype.slice.call(arguments);
   return util.format.apply(util, wrapErrorsWithInspect(data));
 }
 
