@@ -4,7 +4,7 @@ const dgram = require('dgram');
 const util = require('util');
 
 function sendLog(udp, host, port, logObject) {
-  const buffer = new Buffer(JSON.stringify(logObject));
+  const buffer = Buffer.from(JSON.stringify(logObject));
 
   /* eslint no-unused-vars:0 */
   udp.send(buffer, 0, buffer.length, port, host, (err, bytes) => {
@@ -21,6 +21,22 @@ function logstashUDP(config, layout) {
 
   if (!config.fields) {
     config.fields = {};
+  }
+
+  function checkArgs(argsValue, logUnderFields) {
+    if ((!argsValue) || (argsValue === 'both')) {
+      return true;
+    }
+
+    if (logUnderFields && (argsValue === 'fields')) {
+      return true;
+    }
+
+    if ((!logUnderFields) && (argsValue === 'direct')) {
+      return true;
+    }
+
+    return false;
   }
 
   function log(loggingEvent) {
@@ -40,15 +56,17 @@ function logstashUDP(config, layout) {
 
     const fields = {};
     Object.keys(config.fields).forEach((key) => {
-      fields[key] = config.fields[key];
+      fields[key] = typeof config.fields[key] === 'function' ? config.fields[key](loggingEvent) : config.fields[key];
     });
 
     /* eslint no-prototype-builtins:1,no-restricted-syntax:[1, "ForInStatement"] */
     if (loggingEvent.data.length > 1) {
       const secondEvData = loggingEvent.data[1];
-      Object.keys(secondEvData).forEach((key) => {
-        fields[key] = secondEvData[key];
-      });
+      if ((secondEvData !== undefined) && (secondEvData !== null)) {
+        Object.keys(secondEvData).forEach((key) => {
+          fields[key] = secondEvData[key];
+        });
+      }
     }
     fields.level = loggingEvent.level.levelStr;
     fields.category = loggingEvent.categoryName;
@@ -57,13 +75,18 @@ function logstashUDP(config, layout) {
       '@version': '1',
       '@timestamp': (new Date(loggingEvent.startTime)).toISOString(),
       type: type,
-      message: layout(loggingEvent),
-      fields: fields
+      message: layout(loggingEvent)
     };
 
-    Object.keys(fields).forEach((key) => {
-      logObject[key] = fields[key];
-    });
+    if (checkArgs(config.args, true)) {
+      logObject.fields = fields;
+    }
+
+    if (checkArgs(config.args, false)) {
+      Object.keys(fields).forEach((key) => {
+        logObject[key] = fields[key];
+      });
+    }
 
     sendLog(udp, config.host, config.port, logObject);
   }
