@@ -4,6 +4,7 @@
 // load modules
 const log4js = require('log4js');
 const express = require('express');
+
 const app = express();
 
 // config
@@ -25,17 +26,15 @@ const logger = log4js.getLogger('log4jslog');
 // logger.setLevel('ERROR');
 
 // express app
-app.configure(() => {
-  app.use(express.favicon(''));
-	// app.use(log4js.connectLogger(logger, { level: log4js.levels.INFO }));
-	// app.use(log4js.connectLogger(logger, { level: 'auto', format: ':method :url :status' }));
+app.use(express.favicon(''));
+// app.use(log4js.connectLogger(logger, { level: log4js.levels.INFO }));
+// app.use(log4js.connectLogger(logger, { level: 'auto', format: ':method :url :status' }));
 
-	// ### AUTO LEVEL DETECTION
-	// http responses 3xx, level = WARN
-	// http responses 4xx & 5xx, level = ERROR
-	// else.level = INFO
-  app.use(log4js.connectLogger(logger, { level: 'auto' }));
-});
+// ### AUTO LEVEL DETECTION
+// http responses 3xx, level = WARN
+// http responses 4xx & 5xx, level = ERROR
+// else.level = INFO
+app.use(log4js.connectLogger(logger, { level: 'auto' }));
 
 // route
 app.get('/', (req, res) => {
