@@ -40,7 +40,8 @@ function fileAppender(file, layout, logSize, numBackups, options, timezoneOffset
   // there has to be at least one backup if logSize has been specified
   numBackups = numBackups === 0 ? 1 : numBackups;
 
-  debug('Creating file appender (',
+  debug(
+    'Creating file appender (',
     file, ', ',
     logSize, ', ',
     numBackups, ', ',
