@@ -36,11 +36,11 @@ function logFacesAppender(config) {
   return function log(event) {
     // convert to logFaces compact json format
     const lfsEvent = {
-      a: config.application || '',   // application name
-      t: event.startTime.getTime(),  // time stamp
-      p: event.level.levelStr,       // level (priority)
-      g: event.categoryName,         // logger name
-      m: format(event.data)          // message text
+      a: config.application || '', // application name
+      t: event.startTime.getTime(), // time stamp
+      p: event.level.levelStr, // level (priority)
+      g: event.categoryName, // logger name
+      m: format(event.data) // message text
     };
 
     // add context variables if exist
@@ -52,9 +52,7 @@ function logFacesAppender(config) {
     sender.post('', lfsEvent)
       .catch((error) => {
         if (error.response) {
-          console.error(
-            `log4js.logFaces-HTTP Appender error posting to ${config.url}: ${error.response.status} - ${error.response.data}`
-          );
+          console.error(`log4js.logFaces-HTTP Appender error posting to ${config.url}: ${error.response.status} - ${error.response.data}`);
           return;
         }
         console.error(`log4js.logFaces-HTTP Appender error: ${error.message}`);
@@ -67,8 +65,9 @@ function configure(config) {
 }
 
 function format(logData) {
-  const data = Array.isArray(logData) ?
-               logData : Array.prototype.slice.call(arguments);
+  const data = Array.isArray(logData)
+    ? logData
+    : Array.prototype.slice.call(arguments);
   return util.format.apply(util, wrapErrorsWithInspect(data));
 }
 
