@@ -2,6 +2,7 @@
 
 const debug = require('debug')('log4js:multiprocess');
 const net = require('net');
+const CircularJSON = require('circular-json');
 
 const END_MSG = '__LOG4JS__';
 
@@ -19,7 +20,7 @@ function logServer(config, actualAppender, levels) {
     debug('deserialising log event');
     let loggingEvent;
     try {
-      loggingEvent = JSON.parse(msg);
+      loggingEvent = CircularJSON.parse(msg);
       loggingEvent.startTime = new Date(loggingEvent.startTime);
       loggingEvent.level = levels.getLevel(loggingEvent.level.levelStr);
     } catch (e) {
@@ -98,13 +99,13 @@ function workerAppender(config) {
     // The following allows us to serialize errors correctly.
     // Validate that we really are in this case
     const logData = loggingEvent.data.map((e) => {
-      if (e && e.stack && JSON.stringify(e) === '{}') {
+      if (e && e.stack && CircularJSON.stringify(e) === '{}') {
         e = { stack: e.stack };
       }
       return e;
     });
     loggingEvent.data = logData;
-    socket.write(JSON.stringify(loggingEvent), 'utf8');
+    socket.write(CircularJSON.stringify(loggingEvent), 'utf8');
     socket.write(END_MSG, 'utf8');
   }
 
