@@ -113,7 +113,7 @@ function smtpAppender(config, layout, subjectLayout) {
   }
 
   const appender = (loggingEvent) => {
-    unsentCount++;  // eslint-disable-line no-plusplus
+    unsentCount++; // eslint-disable-line no-plusplus
     logEventBuffer.push(loggingEvent);
     if (sendInterval > 0) {
       scheduleSend();
