@@ -8,7 +8,7 @@ const debug = require('debug')('log4js:configuration');
 
 let cluster;
 try {
-  cluster = require('cluster');   // eslint-disable-line global-require
+  cluster = require('cluster'); // eslint-disable-line global-require
 } catch (e) {
   debug('Clustering support disabled because require(cluster) threw an error: ', e);
 }
@@ -36,14 +36,12 @@ function anInteger(thing) {
 }
 
 class Configuration {
-
   throwExceptionIf(checks, message) {
     const tests = Array.isArray(checks) ? checks : [checks];
     tests.forEach((test) => {
       if (test) {
-        throw new Error(
-          `Problem with log4js configuration: (${util.inspect(this.candidate, { depth: 5 })}) - ${message}`
-        );
+        throw new Error(`Problem with log4js configuration: (${util.inspect(this.candidate, { depth: 5 })})` +
+          ` - ${message}`);
       }
     });
   }
