@@ -22,7 +22,7 @@ function getUrl(req) {
 }
 
 
-  /**
+/**
    * Adds custom {token, replacement} objects to defaults,
    * overwriting the defaults if any tokens clash
    *
@@ -37,8 +37,8 @@ function assembleTokens(req, res, customTokens) {
     const a = array.concat();
     for (let i = 0; i < a.length; ++i) {
       for (let j = i + 1; j < a.length; ++j) {
-          // not === because token can be regexp object
-          /* eslint eqeqeq:0 */
+        // not === because token can be regexp object
+        /* eslint eqeqeq:0 */
         if (a[i].token == a[j].token) {
           a.splice(j--, 1);
         }
@@ -91,15 +91,15 @@ function assembleTokens(req, res, customTokens) {
     token: /:res\[([^\]]+)]/g,
     replacement: function (_, field) {
       return res._headers ?
-               (res._headers[field.toLowerCase()] || res.__headers[field])
-          : (res.__headers && res.__headers[field]);
+        (res._headers[field.toLowerCase()] || res.__headers[field])
+        : (res.__headers && res.__headers[field]);
     }
   });
 
   return arrayUniqueTokens(customTokens.concat(defaultTokens));
 }
 
-  /**
+/**
    * Return formatted log line.
    *
    * @param  {String} str
@@ -114,7 +114,7 @@ function format(str, tokens) {
   return str;
 }
 
-  /**
+/**
    * Return RegExp Object about nolog
    *
    * @param  {String|Array} nolog
@@ -154,7 +154,7 @@ function createNoLogCondition(nolog) {
     }
 
     if (Array.isArray(nolog)) {
-        // convert to strings
+      // convert to strings
       const regexpsAsStrings = nolog.map(reg => (reg.source ? reg.source : reg));
       regexp = new RegExp(regexpsAsStrings.join('|'));
     }
