@@ -112,10 +112,10 @@ function timestampLevelAndCategory(loggingEvent, colour, timezoneOffset) {
  */
 function basicLayout(loggingEvent, timezoneOffset) {
   return timestampLevelAndCategory(
-      loggingEvent,
-      undefined,
-      timezoneOffset
-    ) + formatLogData(loggingEvent.data);
+    loggingEvent,
+    undefined,
+    timezoneOffset
+  ) + formatLogData(loggingEvent.data);
 }
 
 /**
@@ -124,10 +124,10 @@ function basicLayout(loggingEvent, timezoneOffset) {
  */
 function colouredLayout(loggingEvent, timezoneOffset) {
   return timestampLevelAndCategory(
-      loggingEvent,
-      loggingEvent.level.colour,
-      timezoneOffset
-    ) + formatLogData(loggingEvent.data);
+    loggingEvent,
+    loggingEvent.level.colour,
+    timezoneOffset
+  ) + formatLogData(loggingEvent.data);
 }
 
 function messagePassThroughLayout(loggingEvent) {
