@@ -6,7 +6,7 @@ const debug = require('debug')('log4js:logger');
 
 let cluster;
 try {
-  cluster = require('cluster');   // eslint-disable-line global-require
+  cluster = require('cluster'); // eslint-disable-line global-require
 } catch (e) {
   debug('Clustering support disabled because require(cluster) threw an error: ', e);
 }
