@@ -24,9 +24,8 @@ function testAppender(label) {
 test('log4js configuration validation', (batch) => {
   batch.test('should give error if config is just plain silly', (t) => {
     [null, undefined, '', ' ', []].forEach((config) => {
-      const expectedError = new Error(
-        `Problem with log4js configuration: (${util.inspect(config)}) - must be an object.`
-      );
+      const expectedError =
+        new Error(`Problem with log4js configuration: (${util.inspect(config)}) - must be an object.`);
       t.throws(
         () => new Configuration(config),
         expectedError
@@ -37,34 +36,32 @@ test('log4js configuration validation', (batch) => {
   });
 
   batch.test('should give error if config is an empty object', (t) => {
-    const expectedError = new Error(
-      'Problem with log4js configuration: ({}) - must have a property "appenders" of type object.'
-    );
+    const expectedError =
+      new Error('Problem with log4js configuration: ({}) - must have a property "appenders" of type object.');
     t.throws(() => new Configuration({}), expectedError);
     t.end();
   });
 
   batch.test('should give error if config has no appenders', (t) => {
-    const expectedError = new Error(
-      'Problem with log4js configuration: ({ categories: {} }) - must have a property "appenders" of type object.'
-    );
+    const expectedError =
+      new Error('Problem with log4js configuration: ({ categories: {} }) ' +
+      '- must have a property "appenders" of type object.');
     t.throws(() => new Configuration({ categories: {} }), expectedError);
     t.end();
   });
 
   batch.test('should give error if config has no categories', (t) => {
-    const expectedError = new Error(
-      'Problem with log4js configuration: ({ appenders: {} }) - must have a property "categories" of type object.'
-    );
+    const expectedError =
+      new Error('Problem with log4js configuration: ({ appenders: {} }) ' +
+      '- must have a property "categories" of type object.');
     t.throws(() => new Configuration({ appenders: {} }), expectedError);
     t.end();
   });
 
   batch.test('should give error if appenders is not an object', (t) => {
-    const error = new Error(
-      'Problem with log4js configuration: ({ appenders: [], categories: [] })' +
-      ' - must have a property "appenders" of type object.'
-    );
+    const error =
+      new Error('Problem with log4js configuration: ({ appenders: [], categories: [] })' +
+      ' - must have a property "appenders" of type object.');
     t.throws(
       () => new Configuration({ appenders: [], categories: [] }),
       error
@@ -73,10 +70,9 @@ test('log4js configuration validation', (batch) => {
   });
 
   batch.test('should give error if appenders are not all valid', (t) => {
-    const error = new Error(
-      'Problem with log4js configuration: ({ appenders: { thing: \'cheese\' }, categories: {} })' +
-      ' - appender "thing" is not valid (must be an object with property "type")'
-    );
+    const error =
+      new Error('Problem with log4js configuration: ({ appenders: { thing: \'cheese\' }, categories: {} })' +
+      ' - appender "thing" is not valid (must be an object with property "type")');
     t.throws(
       () => new Configuration({ appenders: { thing: 'cheese' }, categories: {} }),
       error
@@ -85,10 +81,8 @@ test('log4js configuration validation', (batch) => {
   });
 
   batch.test('should require at least one appender', (t) => {
-    const error = new Error(
-      'Problem with log4js configuration: ({ appenders: {}, categories: {} })' +
-      ' - must define at least one appender.'
-    );
+    const error = new Error('Problem with log4js configuration: ({ appenders: {}, categories: {} })' +
+      ' - must define at least one appender.');
     t.throws(
       () => new Configuration({ appenders: {}, categories: {} }),
       error
@@ -97,11 +91,9 @@ test('log4js configuration validation', (batch) => {
   });
 
   batch.test('should give error if categories are not all valid', (t) => {
-    const error = new Error(
-      'Problem with log4js configuration: ' +
+    const error = new Error('Problem with log4js configuration: ' +
       '({ appenders: { stdout: { type: \'stdout\' } },\n  categories: { thing: \'cheese\' } })' +
-      ' - category "thing" is not valid (must be an object with properties "appenders" and "level")'
-    );
+      ' - category "thing" is not valid (must be an object with properties "appenders" and "level")');
     t.throws(
       () => new Configuration({ appenders: { stdout: { type: 'stdout' } }, categories: { thing: 'cheese' } }),
       error
@@ -110,27 +102,24 @@ test('log4js configuration validation', (batch) => {
   });
 
   batch.test('should give error if default category not defined', (t) => {
-    const error = new Error(
-      'Problem with log4js configuration: ' +
+    const error = new Error('Problem with log4js configuration: ' +
       '({ appenders: { stdout: { type: \'stdout\' } },\n' +
       '  categories: { thing: { appenders: [ \'stdout\' ], level: \'ERROR\' } } })' +
-      ' - must define a "default" category.'
-    );
+      ' - must define a "default" category.');
     t.throws(
       () => new Configuration({
         appenders: { stdout: { type: 'stdout' } },
-        categories: { thing: { appenders: ['stdout'], level: 'ERROR' } } }
-      ),
+        categories: { thing: { appenders: ['stdout'], level: 'ERROR' } }
+      }),
       error
     );
     t.end();
   });
 
   batch.test('should require at least one category', (t) => {
-    const error = new Error(
-      'Problem with log4js configuration: ({ appenders: { stdout: { type: \'stdout\' } }, categories: {} })' +
-      ' - must define at least one category.'
-    );
+    const error =
+      new Error('Problem with log4js configuration: ({ appenders: { stdout: { type: \'stdout\' } }, categories: {} })' +
+      ' - must define at least one category.');
     t.throws(
       () => new Configuration({ appenders: { stdout: { type: 'stdout' } }, categories: {} }),
       error
@@ -139,12 +128,10 @@ test('log4js configuration validation', (batch) => {
   });
 
   batch.test('should give error if category.appenders is not an array', (t) => {
-    const error = new Error(
-      'Problem with log4js configuration: ' +
+    const error = new Error('Problem with log4js configuration: ' +
       '({ appenders: { stdout: { type: \'stdout\' } },\n' +
       '  categories: { thing: { appenders: {}, level: \'ERROR\' } } })' +
-      ' - category "thing" is not valid (appenders must be an array of appender names)'
-    );
+      ' - category "thing" is not valid (appenders must be an array of appender names)');
     t.throws(
       () => new Configuration({
         appenders: { stdout: { type: 'stdout' } },
@@ -156,12 +143,10 @@ test('log4js configuration validation', (batch) => {
   });
 
   batch.test('should give error if category.appenders is empty', (t) => {
-    const error = new Error(
-      'Problem with log4js configuration: ' +
+    const error = new Error('Problem with log4js configuration: ' +
       '({ appenders: { stdout: { type: \'stdout\' } },\n' +
       '  categories: { thing: { appenders: [], level: \'ERROR\' } } })' +
-      ' - category "thing" is not valid (appenders must contain at least one appender name)'
-    );
+      ' - category "thing" is not valid (appenders must contain at least one appender name)');
     t.throws(
       () => new Configuration({
         appenders: { stdout: { type: 'stdout' } },
@@ -173,12 +158,10 @@ test('log4js configuration validation', (batch) => {
   });
 
   batch.test('should give error if categories do not refer to valid appenders', (t) => {
-    const error = new Error(
-      'Problem with log4js configuration: ' +
+    const error = new Error('Problem with log4js configuration: ' +
       '({ appenders: { stdout: { type: \'stdout\' } },\n' +
       '  categories: { thing: { appenders: [ \'cheese\' ], level: \'ERROR\' } } })' +
-      ' - category "thing" is not valid (appender "cheese" is not defined)'
-    );
+      ' - category "thing" is not valid (appender "cheese" is not defined)');
     t.throws(
       () => new Configuration({
         appenders: { stdout: { type: 'stdout' } },
@@ -190,13 +173,11 @@ test('log4js configuration validation', (batch) => {
   });
 
   batch.test('should give error if category level is not valid', (t) => {
-    const error = new Error(
-      'Problem with log4js configuration: ' +
+    const error = new Error('Problem with log4js configuration: ' +
       '({ appenders: { stdout: { type: \'stdout\' } },\n' +
       '  categories: { default: { appenders: [ \'stdout\' ], level: \'Biscuits\' } } })' +
       ' - category "default" is not valid (level "Biscuits" not recognised; ' +
-      'valid levels are ALL, TRACE, DEBUG, INFO, WARN, ERROR, FATAL, MARK, OFF)'
-    );
+      'valid levels are ALL, TRACE, DEBUG, INFO, WARN, ERROR, FATAL, MARK, OFF)');
     t.throws(
       () => new Configuration({
         appenders: { stdout: { type: 'stdout' } },
@@ -208,12 +189,10 @@ test('log4js configuration validation', (batch) => {
   });
 
   batch.test('should give error if appender type cannot be found', (t) => {
-    const error = new Error(
-      'Problem with log4js configuration: ' +
+    const error = new Error('Problem with log4js configuration: ' +
       '({ appenders: { thing: { type: \'cheese\' } },\n' +
       '  categories: { default: { appenders: [ \'thing\' ], level: \'ERROR\' } } })' +
-      ' - appender "thing" is not valid (type "cheese" could not be found)'
-    );
+      ' - appender "thing" is not valid (type "cheese" could not be found)');
     t.throws(
       () => new Configuration({
         appenders: { thing: { type: 'cheese' } },
@@ -278,9 +257,7 @@ test('log4js configuration validation', (batch) => {
     sandboxConfig.requires[
       `${path.join(mainPath, '../../node_modules/tap/node_modules/nyc/bin/cheese')}`
     ] = testAppender('correct');
-    const SandboxedConfiguration = sandbox.require(
-      '../../lib/configuration', sandboxConfig
-    );
+    const SandboxedConfiguration = sandbox.require('../../lib/configuration', sandboxConfig);
 
     const config = new SandboxedConfiguration({
       appenders: { thing: { type: 'cheese' } },
